Add task runner
