+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/anomaly/onboard/BaseDetectionOnboardTask.java
@@ -1,14 +1,8 @@
 package com.linkedin.thirdeye.anomaly.onboard;
 
 import com.google.common.base.Preconditions;
-import com.linkedin.thirdeye.anomaly.task.TaskConstants;
-import org.apache.commons.lang.exception.ExceptionUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public abstract class BaseDetectionOnboardTask implements DetectionOnboardTask {
-  private static final Logger LOG = LoggerFactory.getLogger(BaseDetectionOnboardTask.class);
-
   private final String taskName;
   protected DetectionOnboardTaskContext taskContext = new DetectionOnboardTaskContext();
 
@@ -31,22 +25,5 @@ public abstract class BaseDetectionOnboardTask implements DetectionOnboardTask {
   @Override
   public DetectionOnboardTaskContext getTaskContext() {
     return taskContext;
-  }
-
-  @Override
-  public DetectionOnboardTaskStatus call() throws Exception {
-    DetectionOnboardTaskStatus taskStatus = new DetectionOnboardTaskStatus();
-    taskStatus.setTaskStatus(TaskConstants.TaskStatus.RUNNING);
-
-    try {
-      this.run();
-      taskStatus.setTaskStatus(TaskConstants.TaskStatus.COMPLETED);
-    } catch (Exception e) {
-      taskStatus.setTaskStatus(TaskConstants.TaskStatus.FAILED);
-      taskStatus.setMessage(ExceptionUtils.getStackTrace(e));
-      LOG.error("Error encountered when running task: {}", taskName, e);
-    }
-
-    return taskStatus;
   }
 }
